Add notice for when a user already has API keys

+++ includes/class-edd-api.php
@@ -1405,9 +1405,12 @@ class EDD_API {
 
 		switch( $process ) {
 			case 'generate':
-				$this->generate_api_key( $user_id );
-				delete_transient( 'edd-total-api-keys' );
-				wp_redirect( add_query_arg( 'edd-message', 'api-key-generated', 'edit.php?post_type=download&page=edd-tools&tab=api_keys' ) ); exit();
+				if( $this->generate_api_key( $user_id ) ) {
+					delete_transient( 'edd-total-api-keys' );
+					wp_redirect( add_query_arg( 'edd-message', 'api-key-generated', 'edit.php?post_type=download&page=edd-tools&tab=api_keys' ) ); exit();
+				} else {
+					wp_redirect( add_query_arg( 'edd-message', 'api-key-exists', 'edit.php?post_type=download&page=edd-tools&tab=api_keys' ) ); exit();	
+				}
 				break;
 			case 'regenerate':
 				$this->generate_api_key( $user_id, true );
